Move ProjectedCRSSelector and SelectCRSDialog to separate files
