+++ src/gui/georeferencing_dialog.cpp
@@ -49,6 +49,7 @@
 #include "../util_gui.h"
 #include "../util.h"
 #include "../util/scoped_signals_blocker.h"
+#include "widgets/crs_selector.h"
 
 GeoreferencingDialog::GeoreferencingDialog(MapEditorController* controller, const Georeferencing* initial, bool allow_no_georeferencing)
 : QDialog(controller->getWindow(), Qt::WindowSystemMenuHint | Qt::WindowTitleHint),
@@ -821,305 +822,4 @@ const QCursor& GeoreferencingTool::getCursor() const
 {
 	static auto const cursor = QCursor(QPixmap(":/images/cursor-crosshair.png"), 11, 11); // TODO: custom icon
 	return cursor;
-}
-
-
-
-// ### ProjectedCRSSelector ###
-
-ProjectedCRSSelector::ProjectedCRSSelector(const Georeferencing& georef, QWidget* parent)
- : QWidget(parent)
- , georef(georef)
-{
-	num_custom_items = 0;
-	crs_dropdown = new QComboBox();
-	for (auto&& temp : CRSTemplateRegistry().list())
-	{
-		crs_dropdown->addItem(temp->id(), qVariantFromValue<void*>(const_cast<CRSTemplate*>(temp.get())));
-	}
-	
-	connect(crs_dropdown, SIGNAL(currentIndexChanged(int)), this, SLOT(crsDropdownChanged(int)));
-	
-	layout = NULL;
-	crsDropdownChanged(crs_dropdown->currentIndex());
-}
-
-void ProjectedCRSSelector::addCustomItem(const QString& text, int id)
-{
-	crs_dropdown->insertItem(num_custom_items, text, QVariant(id));
-	if (num_custom_items == 0)
-		crs_dropdown->insertSeparator(1);
-	++num_custom_items;
-}
-
-const CRSTemplate* ProjectedCRSSelector::getSelectedCRSTemplate()
-{
-	QVariant item_data = crs_dropdown->itemData(crs_dropdown->currentIndex());
-	if (!item_data.canConvert<void*>())
-		return NULL;
-	return static_cast<CRSTemplate*>(item_data.value<void*>());
-}
-
-QString ProjectedCRSSelector::getSelectedCRSSpec()
-{
-	QVariant item_data = crs_dropdown->itemData(crs_dropdown->currentIndex());
-	if (!item_data.canConvert<void*>())
-		return QString();
-	
-	CRSTemplate* temp = static_cast<CRSTemplate*>(item_data.value<void*>());
-	QString spec = temp->specificationTemplate();
-	
-	int row = 0;
-	for (auto&& param : temp->parameters())
-	{
-		++row;
-		QWidget* edit_widget = layout->itemAt(row, QFormLayout::FieldRole)->widget();
-		auto spec_values = param->specValues(param->value(edit_widget));
-		for (auto&& value :  spec_values)
-			spec = spec.arg(value);
-	}
-	
-	return spec;
-}
-
-int ProjectedCRSSelector::getSelectedCustomItemId()
-{
-	QVariant item_data = crs_dropdown->itemData(crs_dropdown->currentIndex());
-	if (!item_data.canConvert<int>())
-		return -1;
-	return item_data.toInt();
-}
-
-void ProjectedCRSSelector::selectItem(const CRSTemplate* temp)
-{
-	int index = crs_dropdown->findData(qVariantFromValue<void*>(const_cast<CRSTemplate*>(temp)));
-	blockSignals(true);
-	crs_dropdown->setCurrentIndex(index);
-	blockSignals(false);
-}
-
-void ProjectedCRSSelector::selectCustomItem(int id)
-{
-	int index = crs_dropdown->findData(QVariant(id));
-	blockSignals(true);
-	crs_dropdown->setCurrentIndex(index);
-	blockSignals(false);
-}
-
-int ProjectedCRSSelector::getNumParams()
-{
-	auto temp = getSelectedCRSTemplate();
-	if (temp == NULL)
-		return 0;
-	else
-		return (int)temp->parameters().size();
-}
-
-QString ProjectedCRSSelector::getParam(int i)
-{
-	auto temp = getSelectedCRSTemplate();
-	Q_ASSERT(temp != NULL);
-	Q_ASSERT(i < (int)temp->parameters().size());
-	
-	int widget_index = 3 + 2 * i;
-	return temp->parameters().at(i)->value(layout->itemAt(widget_index)->widget());
-}
-
-void ProjectedCRSSelector::setParam(int i, const QString& value)
-{
-	auto temp = getSelectedCRSTemplate();
-	Q_ASSERT(temp != NULL);
-	Q_ASSERT(i < (int)temp->parameters().size());
-	
-	int widget_index = 3 + 2 * i;
-	QWidget* edit_widget = layout->itemAt(widget_index)->widget();
-	if (!edit_widget->hasFocus())
-	{
-		const QSignalBlocker block(edit_widget);
-		temp->parameters().at(i)->setValue(edit_widget, value);
-	}
-}
-
-const Georeferencing& ProjectedCRSSelector::georeferencing() const
-{
-	return georef;
-}
-
-void ProjectedCRSSelector::crsDropdownChanged(int index)
-{
-	Q_UNUSED(index);
-	if (layout)
-	{
-		for (int i = 2 * layout->rowCount() - 1; i >= 2; --i)
-			delete layout->takeAt(i)->widget();
-		if (layout->count() > 0)
-		{
-			layout->takeAt(1);
-			delete layout->takeAt(0)->widget();
-		}
-		delete layout;
-		layout = NULL;
-	}
-	
-	layout = new QFormLayout();
-	layout->addRow(tr("&Coordinate reference system:"), crs_dropdown);
-	
-	QVariant item_data = crs_dropdown->itemData(crs_dropdown->currentIndex());
-	if (item_data.canConvert<void*>())
-	{
-		CRSTemplate* temp = static_cast<CRSTemplate*>(item_data.value<void*>());
-		for (auto&& param : temp->parameters())
-		{
-			layout->addRow(param->name() + ":", param->createEditor(*this));
-		}
-	}
-	
-	setLayout(layout);
-	
-	emit crsEdited(true);
-}
-
-void ProjectedCRSSelector::crsParameterEdited()
-{
-	emit crsEdited(false);
-}
-
-
-
-// ### SelectCRSDialog ###
-
-SelectCRSDialog::SelectCRSDialog(Map* map, QWidget* parent, bool show_take_from_map,
-								 bool show_local, bool show_geographic, const QString& desc_text)
- : QDialog(parent, Qt::WindowSystemMenuHint | Qt::WindowTitleHint), map(map)
-{
-	Q_ASSERT(map->getGeoreferencing().getState() != Georeferencing::ScaleOnly);
-	
-	setWindowModality(Qt::WindowModal);
-	setWindowTitle(tr("Select coordinate reference system"));
-	
-	QLabel* desc_label = NULL;
-	if (!desc_text.isEmpty())
-		desc_label = new QLabel(desc_text);
-	
-	map_radio = show_take_from_map ? (new QRadioButton(tr("Same as map's"))) : NULL;
-	if (map_radio)
-		map_radio->setChecked(true);
-	
-	local_radio = show_local ? (new QRadioButton(tr("Local"))) : NULL;
-	
-	geographic_radio = show_geographic ? (new QRadioButton(tr("Geographic coordinates (WGS84)"))) : NULL;
-	
-	projected_radio = new QRadioButton(tr("From list"));
-	if (!map_radio)
-		projected_radio->setChecked(true);
-	crs_edit = new ProjectedCRSSelector(map->getGeoreferencing());
-	
-	spec_radio = new QRadioButton(tr("From specification"));
-	
-	crs_spec_edit = new QLineEdit();
-	status_label = new QLabel();
-	
-	button_box = new QDialogButtonBox(QDialogButtonBox::Cancel | QDialogButtonBox::Ok);
-	
-	if (map->getGeoreferencing().isLocal())
-	{
-		if (map_radio)
-			map_radio->setText(map_radio->text() + " " + tr("(local)"));
-		if (geographic_radio)
-			geographic_radio->setEnabled(false);
-		projected_radio->setEnabled(false);
-		spec_radio->setEnabled(false);
-		crs_spec_edit->setEnabled(false);
-	}
-	
-	QHBoxLayout* crs_layout = new QHBoxLayout();
-	crs_layout->addSpacing(16);
-	crs_layout->addWidget(crs_edit);
-	
-	crs_spec_layout = new QFormLayout();
-	crs_spec_layout->addRow(tr("CRS Specification:"), crs_spec_edit);
-	crs_spec_layout->addRow(tr("Status:"), status_label);
-	
-	QVBoxLayout* layout = new QVBoxLayout();
-	if (desc_label)
-	{
-		layout->addWidget(desc_label);
-		layout->addSpacing(16);
-	}
-	if (map_radio)
-		layout->addWidget(map_radio);
-	if (local_radio)
-		layout->addWidget(local_radio);
-	if (geographic_radio)
-		layout->addWidget(geographic_radio);
-	layout->addWidget(projected_radio);
-	layout->addLayout(crs_layout);
-	layout->addWidget(spec_radio);
-	layout->addSpacing(16);
-	layout->addLayout(crs_spec_layout);
-	layout->addSpacing(16);
-	layout->addWidget(button_box);
-	setLayout(layout);
-	
-	if (map_radio)
-		connect(map_radio, SIGNAL(clicked()), this, SLOT(updateWidgets()));
-	if (local_radio)
-		connect(local_radio, SIGNAL(clicked()), this, SLOT(updateWidgets()));
-	if (geographic_radio)
-		connect(geographic_radio, SIGNAL(clicked()), this, SLOT(updateWidgets()));
-	connect(projected_radio, SIGNAL(clicked()), this, SLOT(updateWidgets()));
-	connect(spec_radio, SIGNAL(clicked()), this, SLOT(updateWidgets()));
-	connect(crs_edit, SIGNAL(crsEdited(bool)), this, SLOT(updateWidgets()));
-	connect(crs_spec_edit, SIGNAL(textEdited(QString)), this, SLOT(crsSpecEdited(QString)));
-	connect(button_box, SIGNAL(accepted()), this, SLOT(accept()));
-	connect(button_box, SIGNAL(rejected()), this, SLOT(reject()));
-	
-	updateWidgets();
-}
-
-QString SelectCRSDialog::getCRSSpec() const
-{
-	return crs_spec_edit->text();
-}
-
-void SelectCRSDialog::crsSpecEdited(QString)
-{
-	spec_radio->setChecked(true);
-	
-	updateWidgets();
-}
-
-void SelectCRSDialog::updateWidgets()
-{
-	if (map_radio && map_radio->isChecked())
-		crs_spec_edit->setText(map->getGeoreferencing().isLocal() ? "" : map->getGeoreferencing().getProjectedCRSSpec());
-	else if (local_radio && local_radio->isChecked())
-		crs_spec_edit->setText("");
-	else if (geographic_radio && geographic_radio->isChecked())
-		crs_spec_edit->setText("+proj=latlong +datum=WGS84");
-	else if (projected_radio->isChecked())
-		crs_spec_edit->setText(crs_edit->getSelectedCRSSpec());
-	
-	crs_edit->setEnabled(projected_radio->isChecked());
-	crs_spec_layout->itemAt(0, QFormLayout::LabelRole)->widget()->setVisible(spec_radio->isChecked());
-	crs_spec_edit->setVisible(spec_radio->isChecked());
-	
-	// Update status field and enable/disable ok button
-	bool valid;
-	QString error_text;
-	if (crs_spec_edit->text().isEmpty())
-		valid = true;
-	else
-	{
-		Georeferencing georef;
-		valid = georef.setProjectedCRS("", crs_spec_edit->text());
-		if (!valid)
-			error_text = georef.getErrorText();
-	}
-	
-	button_box->button(QDialogButtonBox::Ok)->setEnabled(valid);
-	if (error_text.length() == 0)
-		status_label->setText(tr("valid"));
-	else
-		status_label->setText(QString("<b>") % error_text % "</b>");
 }
