+++ src/gui/georeferencing_dialog.h
@@ -316,119 +316,4 @@ private:
 
 
 
-/// Combobox for projected coordinate reference system (CRS) selection,
-/// with edit widgets below to specify the free parameters, if necessary.
-class ProjectedCRSSelector : public QWidget, public CRSParameterWidgetObserver
-{
-Q_OBJECT
-public:
-	ProjectedCRSSelector(const Georeferencing& georef, QWidget* parent = NULL);
-	
-	/// Adds a custom text item at the top which can be identified by the given id.
-	void addCustomItem(const QString& text, int id);
-	
-	
-	/// Returns the selected CRS template,
-	/// or NULL if a custom item is selected
-	const CRSTemplate* getSelectedCRSTemplate();
-	
-	/// Returns the selected CRS specification string,
-	/// or an empty string if a custom item is selected
-	QString getSelectedCRSSpec();
-	
-	/// Returns the id of the selected custom item,
-	/// or -1 in case a normal item is selected
-	int getSelectedCustomItemId();
-	
-
-	/// Selects the given item
-	void selectItem(const CRSTemplate* temp);
-	
-	/// Selects the given item
-	void selectCustomItem(int id);
-	
-	
-	/// Returns the number of parameters shown currently
-	int getNumParams();
-	
-	/// Returns the i-th parameters' value (for storage,
-	/// not for pasting into the crs specification!)
-	QString getParam(int i);
-	
-	/// Sets the i-th parameters' value.
-	/// Does not emit crsEdited().
-	void setParam(int i, const QString& value);
-	
-	
-	/**
-	 * Returns the current georeferencing.
-	 */
-	const Georeferencing& georeferencing() const override;
-	
-signals:
-	/// Called when the user edit the CRS.
-	/// system_changed is true if the whole system was switched,
-	/// if only a parameter was changed it is false.
-	void crsEdited(bool system_changed);
-	
-private slots:
-	void crsDropdownChanged(int index);
-	void crsParameterEdited() override;
-	
-private:
-	const Georeferencing& georef;
-	QComboBox* crs_dropdown;
-	int num_custom_items;
-	
-	QLineEdit* param_edit;
-	QSpinBox* param_int_spinbox;
-	
-	QFormLayout* layout;
-};
-
-
-
-/** Dialog to select a coordinate reference system (CRS) */
-class SelectCRSDialog : public QDialog
-{
-Q_OBJECT
-public:
-	/**
-	 * Creates a SelectCRSDialog.
-	 * 
-	 * @param map The map to create the dialog for.
-	 * @param parent The parent widget.
-	 * @param show_take_from_map Toggle whether to show the "Take the map's CRS" option.
-	 * @param show_local Toggle whether to show the "Local" option.
-	 * @param show_geographic Toggle whether to show the "Geographic (WGS84)" option.
-	 * @param desc_text Optional description text for the dialog. Should explain
-	 *                  for what the selected CRS will be used for.
-	 */
-	SelectCRSDialog(Map* map, QWidget* parent, bool show_take_from_map,
-					bool show_local, bool show_geographic, const QString& desc_text = QString());
-	
-	/** Returns the chosen CRS spec after the dialog has completed. */
-	QString getCRSSpec() const;
-	
-private slots:
-	void crsSpecEdited(QString text);
-	void updateWidgets();
-	
-private:
-	/* Internal state */
-	Map* const map;
-	
-	/* GUI elements */
-	QRadioButton* map_radio;
-	QRadioButton* local_radio;
-	QRadioButton* geographic_radio;
-	QRadioButton* projected_radio;
-	QRadioButton* spec_radio;
-	ProjectedCRSSelector* crs_edit;
-	QFormLayout* crs_spec_layout;
-	QLineEdit* crs_spec_edit;
-	QLabel* status_label;
-	QDialogButtonBox* button_box;
-};
-
 #endif
