+++ src/template_image.cpp
@@ -36,6 +36,7 @@
 
 #include "core/georeferencing.h"
 #include "gui/georeferencing_dialog.h"
+#include "gui/select_crs_dialog.h"
 #include "map.h"
 #include "util.h"
